+++ RELEASES.md
@@ -1681,3 +1681,13 @@ Based on:
 - [typescript v0.1.0-alpha.162] .
 ### Releases
 - [NPM v0.1.0-alpha.162] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.162 - .
+
+## 2025-05-27 16:42:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.493.2 (2.512.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.0-alpha.163] .
+### Releases
+- [NPM v0.1.0-alpha.163] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.163 - .